+++ scenario-api/src/main/scala/pl/touk/nussknacker/engine/graph/subprocess.scala
@@ -5,6 +5,7 @@ import pl.touk.nussknacker.engine.graph.evaluatedparam.Parameter
 
 object subprocess {
 
-  @JsonCodec case class SubprocessRef(id: String, parameters: List[Parameter])
+  //TODO: outputVariableNames has to be optional because of backward compatibility (nodes from db are converted to this..), remove optional in next version
+  @JsonCodec case class SubprocessRef(id: String, parameters: List[Parameter], outputVariableNames: Option[Map[String, String]] = None)
 
 }