+++ ui/client/types/node.ts
@@ -32,6 +32,7 @@ export type NodeType<F extends Field = Field> = {
     id: string,
     typ: string,
     parameters: $TodoType[],
+    outputVariableNames: Record<string, string>,
   },
   varName?: string,
   value?: $TodoType,